refactor: Extract depth building
